+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   articles/empirical-comparisons: empirical-comparisons.html
   articles/simulations: simulations.html
-last_built: 2024-11-29T04:12Z
+last_built: 2025-01-14T21:24Z
 urls:
   reference: https://poissonconsulting.github.io/bbousuite/reference
   article: https://poissonconsulting.github.io/bbousuite/articles
