### 2025-01-20 #7650 @kthyng
Hi @ @sebdalgarno 👋  — unfortunately, I am desk rejecting this submission due to having under 300 lines of code. You can see more details in our docs here: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort. Your submission is shown above as having 80 lines of code.
