### 2025-01-20 #7650 @sebdalgarno
hi @kthyng this is a suite of 5 individual R packages. The Github repo that the paper is in is only the wrapper package that installs and loads all of them together.
