### 2025-01-20 #7650 @kthyng
@sebdalgarno Is this like this other submission? https://github.com/openjournals/joss-reviews/issues/7547
