### 2025-01-20 #7650 @sebdalgarno
> [@sebdalgarno](https://github.com/sebdalgarno) Is this like this other submission? [#7547](https://github.com/openjournals/joss-reviews/issues/7547)

Yes different main authors but we work at the same company so the approach is very similar. The project/functionality etc. are different, however.
