### 2025-02-24 #7650 @kthyng
```
>>> cloc bbousuite-main.zip
      78 text files.
      48 unique files.
     149 files ignored.

Language                     files          blank        comment           code
CSV                             18              0              0           1102
TeX                              2             43              0            497
Markdown                         7            106              4            249
YAML                             6             30              9            154
R                               12             16             42             80
Rmd                              3             89            204             59
SUM:                            48            284            259           2141

>>> cloc bboudata-main.zip
      56 text files.
      33 unique files.
      39 files ignored.

Language                     files          blank        comment           code
Markdown                        11            147              3           7950
R                               16            140            607            789
YAML                             5             23              8            138
Rmd                              1             18             36              7
SUM:                            33            328            654           8884

>>> cloc bbouretro-main.zip
      68 text files.
      49 unique files.
      41 files ignored.

Language                     files          blank        comment           code
R                               28            239            750           1666
Markdown                        11            139             17           1575
TeX                              1             27              0            329
YAML                             5             28              8            149
Rmd                              3            107            204             38
CSV                              1              0              0             28
SUM:                            49            540            979           3785

>>> cloc bboutools-main.zip
     254 text files.
     177 unique files.
     122 files ignored.

Language                     files          blank        comment           code
CSV                             85              0              0           5880
R                               69            550           1722           2958
Markdown                         4            143              4            385
Rmd                              4            247            524            242
TeX                              2             18              0            214
YAML                             5             30              8            142
SVG                              8              0              0              8
SUM:                           177            988           2258           9829

>>> cloc bbousims-main.zip
      90 text files.
      58 unique files.
      46 files ignored.

Language                     files          blank        comment           code
Markdown                        13          62092              4         124848
R                               37            261            949           1872
Rmd                              4            167            355            312
YAML                             4             26              8            128
SUM:                            58          62546           1316         127160

>>> cloc bboushiny-main.zip
      52 text files.
      41 unique files.
      15 files ignored.

Language                     files          blank        comment           code
R                               22            477            356           3249
Markdown                        12            281             79            758
YAML                             6             32              8            176
Rmd                              1             30             53              9
SUM:                            41            820            496           4192

```
