### 2025-03-10 #7650 @kthyng
@sebdalgarno Just checking in on you here! How is the paper going?
