+++ paper/paper.R
@@ -1,2 +1,0 @@
-library(bbousuite)
-print(bboudata::bbousurv_a, n = 5)
