+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   articles/empirical-comparisons: empirical-comparisons.html
   articles/simulations: simulations.html
-last_built: 2025-03-10T19:42Z
+last_built: 2025-03-10T19:45Z
 urls:
   reference: https://poissonconsulting.github.io/bbousuite/reference
   article: https://poissonconsulting.github.io/bbousuite/articles
