### 2025-03-17 #7650 @kthyng
@ethanwhite Can you edit this submission? Note that it is an umbrella submission that covers several repositories noted above. Reach out in the JOSS Slack as needed to information on how to organize this, if you are able to take this on.
