### 2025-03-31 #7650 @ethanwhite
Hi @sebdalgarno - I've done an initial check of the paper/software and everything looks good so I'm going to start looking for reviewers. I'm just getting started editing for JOSS, so apologies in advance for any hiccups along the way as I figure things out.
