### 2025-03-31 #7650 @ethanwhite
👋 @BrandonEdwards - would you be willing to review this submission for JOSS? I see you just submitted a review, so totally OK to say no, but I wanted to ask since your experience with the structure of the packages in #7752 is directly relevant to this submission, which is by the same team and uses a similar overarching design.
