### 2025-04-02 #7650 @ethanwhite
@see24 I agree that this doesn't seem like an issue as long as you don't think it would affect your ability to provide an unbiased review. I'm going to tag in the section lead to double check just in case.

@kthyng - I just wanted to check with you that "I am working on a team that is hoping to collaborate with Seb and his team in the future but we have not have any contact so far" doesn't represent a conflict of interest.
