### 2025-04-02 #7650 @BrandonEdwards
> 👋 [@BrandonEdwards](https://github.com/BrandonEdwards) - would you be willing to review this submission for JOSS? I see you just submitted a review, so totally OK to say no, but I wanted to ask since your experience with the structure of the packages in [#7752](https://github.com/openjournals/joss-reviews/issues/7752) is directly relevant to this submission, which is by the same team and uses a similar overarching design.

Hi there,

Sorry that I missed this. I need to change my Github account over to my new email clearly!

It looks like you have reviewers for this now, yes? If something changes, and you still need a reviewer, I can commit to doing a review **after April 18th.** If this is okay, feel free to add me as a reviewer, but otherwise please definitely keep me in mind for future reviews!
