### 2025-04-03 #7650 @kthyng
> "I am working on a team that is hoping to collaborate with Seb and his team in the future but we have not have any contact so far" doesn't represent a conflict of interest.

@ethanwhite This is ok, you can proceed.
