### 2025-04-03 #7997 @editorialbot
[REVIEW]: bbousuite: A set of R packages to facilitate analysis of boreal caribou survival   and recruitment data
**Submitting author:** @sebdalgarno (<a href="http://orcid.org/0000-0002-3658-4517">Sebastian Dalgarno</a>)
**Repository:** https://github.com/poissonconsulting/bbousuite
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
