### 2025-04-04 #7997 @sebdalgarno
thanks @ethanwhite and reviewers - here's a bit more context to make the review more manageable.
bboutools is the workhorse of the suite, containing the bayesian and ML analysis functions. This will likely require the most attention.
bbouretro is a simpler package that focuses on recreating the core bboutools functionality using frequentist methods.
bboudata is a simple package that exists only to load simulated and real (anonymized) datasets. These are mostly used for testing and as demo data in the app.
bbousims was developed to simulate datasets from known rates and compare estimates from the various analytical methods. Although there is some complexity in this package, I would suggest that it does not require as thorough review (i.e., compared to bboutools) given that it is not the primary focus of the paper.
bboushiny is an R package to run a shiny app. We have tried to separate the analysis code from the shiny code as much as possible to make testing and maintenance (and hopefully review!) more straightforward. This contains the shiny code. The analysis code (and related tests) resides in bboutools.
Finally, bbousuite is a very simple package that just loads the other packages (i.e., similar to 'tidyverse').

I hope that helps as you approach your review! Also I hope that it is OK that I provide some input on this.
