### 2025-04-17 #7997 @ethanwhite
👋 @rerickson-usgs, @see24. Thanks again for agreeing to review this paper. I can see that @see24's checklist is already making progress. Much appreciated! Since we're now two weeks into the review process, I just wanted to check in and see if either of you have any questions or have run into any issues that I can help out with as you work on your reviews.
