### 2025-04-17 #7997 @rerickson-usgs
@sebdalgarno I'm starting to review your packages. They seem cool. Right now, I have some questions about data:

1. The example population files (e.g., `bbousurv_a`, `bbourecruit_b`) lack metadata. I understand that these are representative data, but pointing towards a paper, report, or similar resource would help the reader see where the data were collected and provide context.
2. Please include a statement about animal care and use approval or lack of the need for approval (I know that management data in the US often does not require approval, but am not familiar with Canadian law).

I think point 2 would be solved by including some sort of citation or metadata to show you did not collect the data for this project and that it has already been made public elsewhere.

Please let me know how you address these points so that I can check off the data-box.
