### 2025-04-17 #7997 @sebdalgarno
hi @rerickson-usgs thanks for the feedback. We have some basic metadata for each dataset (in bbousurv.R and bbourecruit.R files, and accessible via documentation, i.e., ?bbousurv_a). This includes licensing info for the data. As you note, the data have been anonymized so that they are not traceable to any particular population/study/region.
Let me know if you've seen the existing documentation and if yes, I can try to better address your question.
