### 2025-04-17 #7997 @rerickson-usgs
@sebdalgarno thank you. I look at those examples, e.g.,

> Description
> The data contains survival information for boreal caribou population A which spans 31 years.

But do not think you have enough metadata. Something like:

> Description
> The data contains survival information for boreal caribou population A which spans 31 years. Population A was anonymized based upon empirical data collected by/for [use the correct words here] the Provience of Alberta as part of the endangered species management. Smith et al. (2009) provides an overview/example/representation of how these data might be collected and analyzed.

 I understand you're walking a tight line working with endangered species data. However, simply stating that would help the reader.

Also, overall, I'm really liking your packages. You and your team have done a great job. That's the only thing I've seen to change.
