### 2025-04-17 #7997 @rerickson-usgs
@sebdalgarno  also, this page https://poissonconsulting.github.io/bbousuite/articles/empirical-comparisons.html has the details I want, but my concern is that your package becomes abandoned on a CRAN repo and the website is long gone. The data package will hopefully stand on its own and not depend upon an external website.
