### 2025-04-17 #7997 @sebdalgarno
thanks @rerickson-usgs I can definitely add that the data have been anonymized (i.e., within the dataset documentation). I can also ask my colleagues if they are aware of papers I can cite that describes the general methods for collecting these data (i.e., aerial surveys and radio collaring). I'm hesitant to provide more specific information about the dataset in case that allows identification of the population.
This is a bit of a tricky question as anything more code/model-related I can immediately address but changes regarding how we present the data would have to be approved by our funders (i.e., Alberta Gov). Not to say it can't be done!  Just testing the waters of what you think would be acceptable.
