### 2025-04-17 #7997 @rerickson-usgs
[Note, I wrote this comment before a comment by @sebdalgarno]

@sebdalgarno  and @ethanwhite other than my data question and a question in this post about R-CMD-Checks on GitHub, I've completed my review.

I think this is a well-done set of packages. I did not examine each file, but tried to spot check the packages.

I like how you output the plots as `ggplot2`  objects so that I could do things like:

```{r}
e_plot <- bbr_plot_survival(survival_est)
e_plot + theme(strip.background = element_blank())
```

I noticed that `bbousims` and `bbousims` are failing the R-CMD-Check on GitHub. @ethanwhite not sure if these need to be fixed for JOSS.

I did get a weird warning, but that might simply be my [work] R version not playing nice with anything right now:

```{r}
> ?bbousurv_a
Warning messages:
1: In readLines(f, n) :
  incomplete final line found on '\\wsl.localhost/Ubuntu/home/rerickson/bbousuite/DESCRIPTION'
2: In readLines(f, n) :
```

Once @sebdalgarno follows-up with data and @ethanwhite follow's up with my question, I think my review is done.

@sebdalgarno  very good job. I enjoyed reading your package. Well written and meeting an important need.
