### 2025-04-17 #7997 @rerickson-usgs
> thanks [@rerickson-usgs](https://github.com/rerickson-usgs) I can definitely add that the data have been anonymized (i.e., within the dataset documentation). I can also ask my colleagues if they are aware of papers I can cite that describes the general methods for collecting these data (i.e., aerial surveys and radio collaring). I'm hesitant to provide more specific information about the dataset in case that allows identification of the population. This is a bit of a tricky question as anything more code/model-related I can immediately address but changes regarding how we present the data would have to be approved by our funders (i.e., Alberta Gov). Not to say it can't be done! Just testing the waters of what you think would be acceptable.

I don't want to hangup your project and make you get extra approval for data.

I would be happy with the anonymized part and trying to include a broad example. I know it's hard to find public datasets and apricate that you want/need/legally must protect endangered species data.

If you can find broad methods papers, that would be helpful. Otherwise, simply adding a second sentence like

> Population A was anonymized based upon empirical data collected by/for [use the correct words here] the Provience of Alberta as part of the endangered species management.

would help.

Also, thank you for your quick replies. Real time review is much better than months long back-and-forth!
