### 2025-04-17 #7997 @sebdalgarno
Great thanks @rerickson-usgs I will make the change as you suggest and will reach out to our co-author regarding methods paper citation to include.
I'm also going to look into the R-CMD checks and fix if possible.
