### 2025-04-17 #7997 @sebdalgarno
hi @rerickson-usgs our coauthor John Boulanger recommended that we use DeCesare et al. 2012 paper (Estimating Ungulate Recruitment and Growth Rates Using Age Ratios) for citing data format/collection methods. I will add that!
