### 2025-04-18 #7997 @rerickson-usgs
@sebdalgarno thank you for your replies. The updated documentation looks great. Please ping me with the R-CMD checks question (or @ethanwhite please let me know what JOSS wants for those tests)
