### 2025-04-21 #7997 @ethanwhite
Thanks @rerickson-usgs and @sebdalgarno for the review and responses. I love it when we can work collaboratively in near real-time on things like this. It's not crucial that R-CMD checks be passing, but it looks like @sebdalgarno got them fixed so we're all set either way!

@rerickson-usgs - If I'm following everything correctly your concerns about animal research have been addressed. If so, go ahead and check that last box. If not, can you let @sebdalgarno and I know what's still missing? Thanks!
