### 2025-04-21 #7997 @rerickson-usgs
@ethanwhite  I'm done with my review. I think the authors have meet all of JOSS's guidelines.

Good job @sebdalgarno and thank you for including my suggestions.

Please let me know if you need anything else.
