### 2025-04-24 poissonconsulting/bbousuite#12 @see24
Review of documentation for JOSS publication
I am putting a series of edits and ideas for the articles and readmes of the various packages here because it feels simpler than creating separate issues in each repo. Hopefully that works for you!

## bbousuite

- [x] Could you include the context you provided in the JOSS review issue about what each package is for in the README?

- [x] Also a list of the vignettes with a short description would be helpful.

This could help answer where should a new user go first? For example I started reading the empirical comparison article first but found that it seemed the best place to start was the bboutools readme and then analytical methods article.

- [x] "Copyright xx xx under licensing" should this placeholder be removed?

### Simulations article

- [x] Figure 1 could you make the stroke width larger for the lines in the legend? It is hard to see the colours.

- [x] In the discussion you say that for recruitment Bayesian random effects performed best but I don't think that is held up by the plots. It looks like all models perform similarly for Bias, Bayesian Random has the narrowest CIs but Retro has the best coverage. I think the conclusion that Bayesian Random performs best for growth still makes sense but would need a little more nuance to align with the data.

## bbouretro
### Traditional methods article

- [x] It seems like the headings are weird in the TOC, especially lambda but I also don't know why S and R are repeated 3 times.

### Methods used in bbouretro article

- [x] Contains this vignette code but it is not a link and doesn't seem to be one of the vignettes `vignette("previous-methods")`. Should it be `vignette("traditional-methods")`

## bboutools
### Priors selection article

- [x] Citations are not working properly.

- [x] There are several prompts to see a vignette or article for details but these are not linked. It would be nice if they were links and I would pick either vignette or article as the terminology to refer to them. (I think this also applies to the Get started page)

### Get started

- [x] Typo in Population growth section `bb_fir_survival` instead of `bb_fit_survival`

## bbousims

### General tools article

- [x] Typo in first sentence "todo" should be "to do"

## bboushiny

- [x] In Overview of how to use the app "Download the template for the tab" it is not clear what is meant by tab maybe "Download the data templates for survival and recruitment"

### In the app
#### About tab

- [x] Typo in last sentence of recruitment section "The annual predicted rescruitment"

#### Help tab

These two sentences are confusing:

- [x] "For descriptions on each columns check the instructions on the tab. " maybe something like "Columns are described in the Instructions pop up above the Download template button "

- [x] "Each tab has its own Instructions help file which is a located in the Instructions box next to the word Instructions. To have the instructions pop up press the circle with the question mark in it." Change to something like "Each of the Survival and Recruitment tabs have an Instructions file that can be accessed by clicking the circle with the question mark in it beside the word Instructions at the top of the panel."
