### 2025-04-29 2b77b33
add more detail to readme for individual pkg function
