### 2025-04-29 3783676
add links to vignettes
