+++ R/licensing.R
@@ -7,6 +7,7 @@
 #' licensing_md()
 licensing_md <- function() {
   "## Licensing
+  Copyright 2022-2023 Environment and Climate Change Canada\\
   Copyright 2023-2024 Province of Alberta\\
 
   The documentation is released under the
