### 2025-05-04 poissonconsulting/bbousuite#12 @sebdalgarno
thanks @see24  for the thorough documentation review!
I've fixed everything you've suggested, except the item for bboutools Readme "There are several prompts to see a vignette or article for details but these are not linked...". From what I can tell all of the articles/repos etc. referenced are hyperlinks.
I've opened PR in each repo with changes.
Can you confirm that the hyperlinks don't work?
