+++ R/devtools-helpers.R
@@ -1,4 +1,6 @@
-# Copyright 2023 Province of Alberta
+# Copyright 2022 Environment and Climate Change Canada
+# Copyright 2023 Environment and Climate Change Canada
+# Copyright 2024 Province of Alberta
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
