+++ R/licensing.R
@@ -7,8 +7,9 @@
 #' licensing_md()
 licensing_md <- function() {
   "## Licensing
-  Copyright 2022-2023 Environment and Climate Change Canada\\
-  Copyright 2023-2024 Province of Alberta\\
+  Copyright 2022 Environment and Climate Change Canada\\
+  Copyright 2023 Environment and Climate Change Canada\\
+  Copyright 2024 Province of Alberta\\
 
   The documentation is released under the
   [CC BY 4.0 License](https://creativecommons.org/licenses/by/4.0/)
