### 2025-05-04 e2a568d
add licensing comments
