+++ R/licensing.R
@@ -1,3 +1,19 @@
+# Copyright 2022 Environment and Climate Change Canada
+# Copyright 2023 Environment and Climate Change Canada
+# Copyright 2024 Province of Alberta
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 #' Licensing Markdown
 #'
 #' A string of markdown code indicating the licensing of the code
