### 2025-05-04 ef5d983
fix copyrigh tin licensing function
