+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   articles/empirical-comparisons: empirical-comparisons.html
   articles/simulations: simulations.html
-last_built: 2025-03-10T19:45Z
+last_built: 2025-05-04T21:05Z
 urls:
   reference: https://poissonconsulting.github.io/bbousuite/reference
   article: https://poissonconsulting.github.io/bbousuite/articles
