### 2025-05-05 #7997 @ethanwhite
Thanks @sebdalgarno and @see24! I really appreciate the two of you working together to further improve this package!

@sebdalgarno - I have it on my To Do list for this week to check everything over and let you know if there are any last little things we need prior to officially accepting the paper.
