### 2025-05-09 #7997 @ethanwhite
Everything looks really good @sebdalgarno. I have just a couple of small asks for minor changes to the paper:

1. In the paper there is a "here" link - "See additional details on bbouretro methods [here](https://poissonconsulting.github.io/bbouretro/articles/retro-methods.html#population-growth-lambda)." It's generally good accessibility practice to have clear indication of what the link is in the link text itself, so can you update this to "See additional details on the [bbouretro methods page](https://poissonconsulting.github.io/bbouretro/articles/retro-methods.html#population-growth-lambda)."
2. Same as above but change "can be found here and here, respectively" to "can be found on the bbousuite [empirical comparison](https://poissonconsulting.github.io/bbousuite/articles/empirical-comparisons.html) and [simulation analysis](https://poissonconsulting.github.io/bbousuite/articles/simulations.html)" pages.
3. Same for "More discussion on the benefits of random effects models can be found here." -> "More discussion on the benefits of random effects models can be found in the [Discussion section of the simulation analysis page](https://poissonconsulting.github.io/bbousuite/articles/simulations.html#discussion).

Please make those changes, then make a new release of the software, and then create a new archive of that release on Zenodo/figshare and provide a link to the release on GitHub and the associated archive. Some of these steps will be listed on the post review checklist that I'll create in a minute.
