### 2025-05-14 #7997 @sebdalgarno
thanks @ethanwhite.
The version number is v0.1.1.
GitHub release is here: https://github.com/poissonconsulting/bbousuite/releases/tag/v0.1.1
the DOI is: 10.5281/zenodo.15400142
