### 2025-05-14 #7997 @ethanwhite
Thanks @sebdalgarno! I've completed all of my checks and everything looks great. I've recommended acceptance and now it's on the EIC's To Do list to make final checks and push the accept button.
