### 2025-05-23 #7997 @kthyng
@sebdalgarno Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "canada"
