+++ paper/paper.bib
@@ -110,7 +110,7 @@
 }
 
 @article{hatter_moose_1991,
-	title = {Moose {Recuriment} {Adult} {Mortality} and {Rate} of {Change}},
+	title = {Moose {Recruitment}, {Adult} {Mortality} and {Rate} of {Change}},
 	volume = {27},
 	author = {Hatter, Ian and Bergerud, Wendy},
 	year = {1991},
@@ -162,7 +162,7 @@
   title={Caribou in Canada}, 
   url={https://www.canada.ca/en/environment-climate-change/services/species-risk-education-centre/caribou.html}, 
   journal={Government of Canada}, 
-  author={Environment & Climate Change Canada, I.}, 
+  author={Environment & Climate Change Canada}, 
   year={2023}, 
   month={Feb},
   urldate = {2024-07-17},
