+++ paper/paper.bib
@@ -112,6 +112,7 @@
 @article{hatter_moose_1991,
 	title = {Moose {Recruitment}, {Adult} {Mortality} and {Rate} of {Change}},
 	volume = {27},
+	journal = {Alces},
 	author = {Hatter, Ian and Bergerud, Wendy},
 	year = {1991},
 	keywords = {boreal-caribou},
