### 2025-05-26 61db259
fix smith 2004 reference
