+++ paper/paper.bib
@@ -1,13 +1,12 @@
 @article{Hatter_2020, title={Revisiting the recruitment-mortality equation to assess moose growth rates}, volume={56}, url={https://alcesjournal.org/index.php/alces/article/view/263}, abstractNote={Hatter and Bergerud (1991) developed a recruitment-mortality (R-M) equation to estimate the annual finite rate of change (λ) in a moose (&amp;lt;em&amp;gt;Alces alces&amp;lt;/em&amp;gt;) population from a single estimate of calf recruitment and adult mortality. I present and assess an alternative formulation of the R-M equation and compare it with the original. A modification to the R-M equations is provided to accommodate early to mid-winter composition surveys where recruitment is measured when calves are less than 1 year-of-age. An example with the modified R-M equation illustrates estimation of λ for the female component of two moose populations under recent study in British Columbia, Canada. Due to potential biases with estimating recruitment and mortality rates, the calculation of λ with the R-M equation should be verified with periodic density surveys whenever possible. The R-M equation is most useful for estimating λ when moose density surveys are not feasible or an estimate of the adult survival rate is available.}, journal={Alces: A Journal Devoted to the Biology and Management of Moose}, author={Hatter, Ian}, year={2020}, month={Aug.}, pages={39–47} }
 
 @article{smith_2004,
-	title = {Woodland {Caribou} {Demography} and {Persistence} {Relative} to {Landscape} {Change} in {West}-central {Alberta}.},
-	language = {en},
-	author = {Smith, Kirby Gordon},
-	year = {2004},
-	keywords = {boreal-caribou},
-	pages = {125},
-	file = {Smith - Woodland Caribou Demography and Persistence Relati.pdf:/Users/sebastiandalgarno/Zotero/storage/D92HPG57/Smith - Woodland Caribou Demography and Persistence Relati.pdf:application/pdf},
+	 title={Woodland Caribou Demography and Persistence Relative to Landscape Change in West-Central Alberta},
+  author={Smith, Kirby Gordon},
+  year={2004},
+  school={University of Alberta},
+  address={Edmonton, AB, Canada},
+  type={M.Sc. Thesis}
 }
 
 
