+++ paper/paper.bib
@@ -162,7 +162,7 @@
   title={Caribou in Canada}, 
   url={https://www.canada.ca/en/environment-climate-change/services/species-risk-education-centre/caribou.html}, 
   journal={Government of Canada}, 
-  author={Environment & Climate Change Canada}, 
+  author={{Environment and Climate Change Canada}}, 
   year={2023}, 
   month={Feb},
   urldate = {2024-07-17},
