+++ paper/paper.bib
@@ -140,7 +140,7 @@
 	issn = {0022541X},
 	url = {https://onlinelibrary.wiley.com/doi/10.1002/jwmg.244},
 	doi = {10.1002/jwmg.244},
-	abstract = {Trends in population grow th can be m onitored w ith data for key vital rates w ithout knowledge o f abundance. A lthough adult female survival has the highest elasticity for ungulate population dynamics, the m ore variable recruitm ent rates are com m only m onitored to track local variation in grow th rates. Specifically, recruitm ent is often m easured using late w in ter youngradult age ratios, th o u g h these age ratios are difficult to reliably interpret given the contribution o f m ultiple vital rates to annual ratios. W e show th at the supplem entation o f age ratio data w ith concurrent radio-telem etry m onitoring o f adult female survival allows both retrospective estim ation o f empirical population grow th rates and the decom position o f recruitm ent-specific vital rates. W e dem onstrate the estim ation o f recruitm ent and population grow th rates for 1 w oodland caribou population using these m ethods, including elasticity and life-stage sim ulation analysis o f the relative contribution o f adult female survival and recruitm ent rates to variation in population growth. W e show, for this w oodland caribou population, th at adult female survival and recruitm ent rates w ere nearly equivalent drivers o f population growth. W e recom m end the concurrent m onitoring o f adult female survival to reliably interpret age ratios w hen m anaging caribou and other ungulates. © 2011 T h e W ildlife Society.},
+	abstract = {Trends in population grow th can be m onitored with data for key vital rates w ithout knowledge o f abundance. A lthough adult female survival has the highest elasticity for ungulate population dynamics, the m ore variable recruitm ent rates are com m only m onitored to track local variation in grow th rates. Specifically, recruitm ent is often m easured using late w in ter youngradult age ratios, th o u g h these age ratios are difficult to reliably interpret given the contribution o f m ultiple vital rates to annual ratios. W e show th at the supplem entation o f age ratio data w ith concurrent radio-telem etry m onitoring o f adult female survival allows both retrospective estim ation o f empirical population grow th rates and the decom position o f recruitm ent-specific vital rates. W e dem onstrate the estim ation o f recruitm ent and population grow th rates for 1 w oodland caribou population using these m ethods, including elasticity and life-stage sim ulation analysis o f the relative contribution o f adult female survival and recruitm ent rates to variation in population growth. W e show, for this w oodland caribou population, th at adult female survival and recruitm ent rates w ere nearly equivalent drivers o f population growth. W e recom m end the concurrent m onitoring o f adult female survival to reliably interpret age ratios w hen m anaging caribou and other ungulates. © 2011 T h e W ildlife Society.},
 	language = {en},
 	number = {1},
 	urldate = {2021-10-11},
@@ -154,7 +154,7 @@
 }
 
 @misc{ECCC_2023, 
-  title={Caribou in Canada}, 
+  title={{Caribou} in {Canada}}, 
   url={https://www.canada.ca/en/environment-climate-change/services/species-risk-education-centre/caribou.html}, 
   journal={Government of Canada}, 
   author={{Environment and Climate Change Canada}}, 
