### 2025-05-26 d52a78a
fix capitalization issues - all sentence case
