+++ paper/paper.bib
@@ -33,7 +33,7 @@
 
 @book{kery_bayesian_2011,
 	address = {Boston},
-	title = {Bayesian population analysis using {WinBUGS}: a hierarchical perspective},
+  title = {Bayesian population analysis using {WinBUGS}: a hierarchical perspective},
 	isbn = {978-0-12-387020-9 0-12-387020-8},
 	shorttitle = {Bayesian population analysis using {WinBUGS}},
 	url = {http://www.vogelwarte.ch/bpa.html},
@@ -45,7 +45,7 @@
 }
 
 @article{de_valpine_programming_2017,
-	title = {Programming {With} {Models}: {Writing} {Statistical} {Algorithms} for {General} {Model} {Structures} {With} {NIMBLE}},
+  title = {Programming with models: writing statistical algorithms for general model structures with {NIMBLE}},
 	volume = {26},
 	issn = {1061-8600},
 	shorttitle = {Programming {With} {Models}},
@@ -66,7 +66,7 @@
 }
 
 @article{eacker_webbased_2019,
-	title = {Web‐based application for threatened woodland caribou population modeling},
+  title = {Web-based application for threatened woodland caribou population modeling},
 	volume = {43},
 	issn = {1938-5463, 1938-5463},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/wsb.950},
@@ -115,7 +115,7 @@
 }
 
 @article{pollock_survival_1989,
-	title = {Survival {Analysis} in {Telemetry} {Studies}: {The} {Staggered} {Entry} {Design}},
+  title = {Survival analysis in telemetry studies: the staggered entry design},
 	volume = {53},
 	issn = {0022541X},
 	shorttitle = {Survival {Analysis} in {Telemetry} {Studies}},
@@ -135,7 +135,7 @@
 }
 
 @article{decesare_estimating_2012,
-	title = {Estimating ungulate recruitment and growth rates using age ratios},
+  title = {Estimating ungulate recruitment and growth rates using age ratios},
 	volume = {76},
 	issn = {0022541X},
 	url = {https://onlinelibrary.wiley.com/doi/10.1002/jwmg.244},
