### 2025-05-26 3465ee7
remove shorttitles and fix more sentence case titles
