+++ paper/paper.bib
@@ -1,4 +1,13 @@
-@article{Hatter_2020, title={Revisiting the recruitment-mortality equation to assess moose growth rates}, volume={56}, url={https://alcesjournal.org/index.php/alces/article/view/263}, abstractNote={Hatter and Bergerud (1991) developed a recruitment-mortality (R-M) equation to estimate the annual finite rate of change (λ) in a moose (&amp;lt;em&amp;gt;Alces alces&amp;lt;/em&amp;gt;) population from a single estimate of calf recruitment and adult mortality. I present and assess an alternative formulation of the R-M equation and compare it with the original. A modification to the R-M equations is provided to accommodate early to mid-winter composition surveys where recruitment is measured when calves are less than 1 year-of-age. An example with the modified R-M equation illustrates estimation of λ for the female component of two moose populations under recent study in British Columbia, Canada. Due to potential biases with estimating recruitment and mortality rates, the calculation of λ with the R-M equation should be verified with periodic density surveys whenever possible. The R-M equation is most useful for estimating λ when moose density surveys are not feasible or an estimate of the adult survival rate is available.}, journal={Alces: A Journal Devoted to the Biology and Management of Moose}, author={Hatter, Ian}, year={2020}, month={Aug.}, pages={39–47} }
+@article{Hatter_2020, 
+  title={Revisiting the recruitment-mortality equation to assess moose growth rates}, 
+  volume={56}, 
+  url={https://alcesjournal.org/index.php/alces/article/view/263}, 
+  journal={Alces: A Journal Devoted to the Biology and Management of Moose},  # ADD THIS
+  author={Hatter, Ian}, 
+  year={2020}, 
+  month={Aug.}, 
+  pages={39–47}
+}
 
 @mastersthesis{smith_2004,
     title={Woodland Caribou Demography and Persistence Relative to Landscape Change in West-Central Alberta},
@@ -11,7 +20,7 @@
 
 @misc{r_core_team_2024,
 	address = {Vienna, Austria},
-	title = {R: {A} {Language} and {Environment} for {Statistical} {Computing}},
+  title = {{R}: a language and environment for statistical computing},
 	url = {https://www.R-project.org/},
 	publisher = {R Foundation for Statistical Computing},
 	author = {{R Core Team}},
@@ -35,7 +44,6 @@
 	address = {Boston},
   title = {Bayesian population analysis using {WinBUGS}: a hierarchical perspective},
 	isbn = {978-0-12-387020-9 0-12-387020-8},
-	shorttitle = {Bayesian population analysis using {WinBUGS}},
 	url = {http://www.vogelwarte.ch/bpa.html},
 	language = {English},
 	publisher = {Academic Press},
@@ -48,7 +56,6 @@
   title = {Programming with models: writing statistical algorithms for general model structures with {NIMBLE}},
 	volume = {26},
 	issn = {1061-8600},
-	shorttitle = {Programming {With} {Models}},
 	url = {https://doi.org/10.1080/10618600.2016.1172487},
 	doi = {10.1080/10618600.2016.1172487},
[…]
