### 2025-05-26 c715375
fix smith thesis sentence case
