+++ paper/paper.bib
@@ -10,8 +10,7 @@
 }
 
 @mastersthesis{smith_2004,
-    title={Woodland Caribou Demography and Persistence Relative to Landscape Change in West-Central Alberta},
-    author={Smith, Kirby Gordon},
+    title={Woodland caribou demography and persistence relative to landscape change in west-central {Alberta}},    author={Smith, Kirby Gordon},
     year={2004},
     school={University of Alberta},
     address={Edmonton, AB, Canada},
