+++ paper/paper.bib
@@ -2,7 +2,7 @@
   title={Revisiting the recruitment-mortality equation to assess moose growth rates}, 
   volume={56}, 
   url={https://alcesjournal.org/index.php/alces/article/view/263}, 
-  journal={Alces: A Journal Devoted to the Biology and Management of Moose},  # ADD THIS
+  journal={Alces: A Journal Devoted to the Biology and Management of Moose}, 
   author={Hatter, Ian}, 
   year={2020}, 
   month={Aug.}, 
